House Price Prediction/requirements.txt
altair==5.0.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.0
contourpy==1.1.0
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.11.0
cymem==2.0.7
cytoolz==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
docx2txt==0.8
easygui==0.98.3
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.2.1
executing==2.0.1
Faker==19.1.0
fastembed==0.2.6
favicon==0.7.0
filelock==3.13.4
flatbuffers==24.3.25
floret==0.10.3
fonttools==4.43.0
fsspec==2024.3.1
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.41
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
htbuilder==0.6.2
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
jellyfish==1.0.0
Jinja2==3.1.3
joblib==1.3.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.4
langcodes==3.3.0
lightgbm==4.4.0
loguru==0.7.2
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
more-itertools==9.1.0
mpmath==1.3.0
murmurhash==1.0.9
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.25.1
onnx==1.16.0
onnxruntime==1.17.3
oscrypto==1.3.0
packaging==23.1
pafy==0.5.5
pandas==2.0.3
parso==0.8.4
pathvalidate==3.2.0
pathy==0.10.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfminer3==2018.12.3.0
pdfplumber==0.11.0
pillow==10.2.0
plac==1.1.3
platformdirs==4.2.1
plotly==5.15.0
preshed==3.0.8
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.15
pydeck==0.8.1b0
Pygments==2.15.1
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pymdown-extensions==10.1
Pympler==1.0.1
PyMySQL==1.1.0
pyparsing==3.0.9
pypdf==3.17.0
pypdfium2==4.29.0
pyphen==0.14.0
pypng==0.20220715.0
pyreadline3==3.4.1
pyresparser==1.0.6
pyrsistent==0.20.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.29.1
regex==2023.6.3
reportlab==3.6.13
requests==2.31.0
rich==13.4.2
rpds-py==0.8.11
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
st-annotated-text==4.0.0
stack-data==0.6.3
streamlit==1.30.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.2
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.5
streamlit-keyup==0.2.0
streamlit-tags==1.2.8
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
svglib==1.5.1
sympy==1.12
tbb==2021.12.0
tenacity==8.2.2
textacy==0.13.0
thinc==8.1.10
threadpoolctl==3.2.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.0
torch==2.3.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
uvicorn==0.23.2
validators==0.20.0
vboxapi==1.0
virtualenv==20.26.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
win32-setctime==1.1.0
xgboost==2.0.3
xhtml2pdf==0.2.11
youtube-dl==2021.12.17
zipp==3.16.2
